Fix SoundCloud HLS stream session timeout
SoundCloud HLS stream segments contain session keys, which become
invalid after 30 minutes. Because the underlying FFmpeg player doesn't
reload the HLS playlist when this happens, the playback stops after
30 minutes. The SoundCloud website addresses this issue by simply
downloading the m3u8-playlist again if a 403 Forbidden HTTP status code
is encountered, thus obtaining new session keys.

The mp3 progressive encoding seems to be not affected by a session
limit. That's why this audio format is now the new default, even if
it is inferior compared to the previous Opus codec.

Resolves #19