Make it possible to use a custom client ID
Until now, the client ID was always overwritten when
Kodi was restarted. With this change, a custom client ID
will not be overwritten anymore.

The service script was removed and replaced with a
getter function which caches the client ID.

The setting `apiv2.clientid` had to be renamed,
because otherwise the plugin would think a custom
client ID was provided. Why? For existing installs
the "custom" client ID was actually set by the plugin
itself. This would lead to a non-updating client ID,
which we want to avoid. By renaming the setting, we
create a "fresh" setting and can work around this
bad design decision.

Resolves #38