Ensure consistent string-encoding for client ID
When getting a new client ID, the client ID was returned as a unicode
string instead of an ASCII string. But the client ID was saved as an
ASCII string in the settings.

Since the client ID is part of the request payload and the caching
mechanism, the encoding difference between a new client ID and a cached
client ID resulted in a cache miss (affecting the "discover"-endpoint).
This resulted in weird behaviour and an error in some cases.
By ensuring that the client ID is always an ASCII string we can avoid
this issue.

Fixes #55